1. Make NA's logical instead of strings in TFoutputResultsFormat()

+++ exec/writeExcel.py
@@ -15,7 +15,7 @@ import operator
 def writeExcel(pandas_df, data_colnames_wanted_py, chart_references_py, file_name_py):
   pandas_df['Stem Group ID'] = pandas_df['Stem Group ID'].fillna(0)
   pandas_df['Banner Group ID'] = pandas_df['Banner Group ID'].fillna(0)
-  chart_references_py['Stem Group ID'] = chart_references_py['Stem Group ID'].fillna(0).astype(int)
+  chart_references_py['Stem Group ID'] = chart_references_py['Stem Group ID'].fillna(0)
   chart_references_py['Banner Group ID'] = chart_references_py['Banner Group ID'].fillna(0)
   
   first_table_row = 0
